Move tests for bulk upload to separate file
